+++ packages/mangrove.js/src/market.ts
@@ -340,8 +340,9 @@ class Market {
   /**
    * Market buy order. Will attempt to buy base token using quote tokens.
    * Params can be of the form:
-   * - `{volume,price}`: buy `wants` tokens for a max average price of `price`, or
-   * - `{wants,gives}`: accept implicit max average price of `gives/wants`
+   * - `{volume,price}`: buy `volume` base tokens for a max average price of `price`. Set `price` to null for a true market order. `fillWants` will be true.
+   * - `{total,price}` : buy as many base tokens as possible using up to `total` quote tokens, with a max average price of `price`. Set `price` to null for a true market order. `fillWants` will be false.
+   * - `{wants,gives,fillWants?}`: accept implicit max average price of `gives/wants`
    *
    * In addition, `slippage` defines an allowed slippage in % of the amount of quote token.
    *
@@ -357,9 +358,25 @@ class Market {
    * ```
    */
   buy(params: TradeParams): Promise<Market.OrderResult> {
-    const _wants = "price" in params ? Big(params.volume) : Big(params.wants);
-    let _gives =
-      "price" in params ? _wants.mul(params.price) : Big(params.gives);
+    let _wants, _gives, fillWants;
+    if ("price" in params) {
+      if ("volume" in params) {
+        _wants = Big(params.volume);
+        _gives =
+          params.price === null
+            ? Big(2).pow(256).minus(1)
+            : _wants.mul(params.price);
+        fillWants = true;
+      } else {
+        _gives = Big(params.total);
+        _wants = params.price === null ? 0 : _gives.div(params.price);
+        fillWants = false;
+      }
+    } else {
+      _wants = Big(params.wants);
+      _gives = Big(params.gives);
+      fillWants = "fillWants" in params ? params.fillWants : true;
+    }
 
     const slippage = validateSlippage(params.slippage);
 
@@ -368,14 +385,15 @@ class Market {
     const wants = this.base.toUnits(_wants);
     const gives = this.quote.toUnits(_gives);
 
-    return this.#marketOrder({ gives, wants, orderType: "buy" });
+    return this.#marketOrder({ gives, wants, orderType: "buy", fillWants });
   }
 
   /**
    * Market sell order. Will attempt to sell base token for quote tokens.
    * Params can be of the form:
-   * - `{volume,price}`: sell `gives` tokens for a min average of `price`
-   * - `{wants,gives}`: accept implicit min average price of `gives/wants`.
+   * - `{volume,price}`: sell `volume` base tokens for a min average price of `price`. Set `price` to null for a true market order. `fillWants` will be false.
+   * - `{total,price}` : sell as many base tokens as possible buying up to `total` quote tokens, with a min average price of `price`. Set `price` to null. `fillWants` will be true.
+   * - `{wants,gives,fillWants?}`: accept implicit min average price of `gives/wants`. `fillWants` will be false by default.
    *
    * In addition, `slippage` defines an allowed slippage in % of the amount of quote token.
    *
@@ -391,9 +409,25 @@ class Market {
    * ```
    */
   sell(params: TradeParams): Promise<Market.OrderResult> {
-    const _gives = "price" in params ? Big(params.volume) : Big(params.gives);
-    let _wants =
-      "price" in params ? _gives.mul(params.price) : Big(params.wants);
+    let _wants, _gives, fillWants;
+    if ("price" in params) {
+      if ("volume" in params) {
+        _gives = Big(params.volume);
+        _wants = params.price === null ? 0 : _gives.mul(params.price);
+        fillWants = false;
+      } else {
+        _wants = Big(params.total);
+        _gives =
+          params.price === null
+            ? Big(2).pow(256).minus(1)
+            : _wants.div(params.price);
+        fillWants = true;
+      }
+    } else {
+      _wants = Big(params.wants);
+      _gives = Big(params.gives);
+      fillWants = "fillWants" in params ? params.fillWants : false;
+    }
 
     const slippage = validateSlippage(params.slippage);
 
@@ -402,16 +436,16 @@ class Market {
     const gives = this.base.toUnits(_gives);
     const wants = this.quote.toUnits(_wants);
 
-    return this.#marketOrder({ wants, gives, orderType: "sell" });
+    return this.#marketOrder({ wants, gives, orderType: "sell", fillWants });
   }
 
   /**
    * Low level Mangrove market order.
    * If `orderType` is `"buy"`, the base/quote market will be used,
-   * with contract function argument `fillWants` set to true.
    *
    * If `orderType` is `"sell"`, the quote/base market will be used,
-   * with contract function argument `fillWants` set to false.
+   *
+   * `fillWants` defines whether the market order stops immediately once `wants` tokens have been purchased or whether it tries to keep going until `gives` tokens have been spent.
    *
    * In addition, `slippage` defines an allowed slippage in % of the amount of quote token.
    *
@@ -422,12 +456,14 @@ class Market {
     wants,
     gives,
     orderType,
+    fillWants,
   }: {
     wants: ethers.BigNumber;
     gives: ethers.BigNumber;
     orderType: "buy" | "sell";
+    fillWants: boolean;
   }): Promise<Market.OrderResult> {
-    const [outboundTkn, inboundTkn, fillWants] =
+    const [outboundTkn, inboundTkn] =
       orderType === "buy"
         ? [this.base, this.quote, true]
         : [this.quote, this.base, false];