+++ packages/mangrove.js/src/types/index.ts
@@ -33,6 +33,7 @@ export interface MarketParams {
 export type Bigish = Big | number | string;
 
 export type TradeParams = { slippage?: number } & (
-  | { volume: Bigish; price: Bigish }
-  | { wants: Bigish; gives: Bigish }
+  | { volume: Bigish; price: Bigish | null }
+  | { total: Bigish; price: Bigish | null }
+  | { wants: Bigish; gives: Bigish; fillWants?: boolean }
 );